chore : rename path in pages
