+++ src/pages/landing/GameSupport.tsx
