+++ src/pages/landing/index.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
 
-import Header from 'pages/LandingPage/Header';
+import Header from 'pages/landing/Header';
 import Footer from 'components/Footer';
-import Content from 'pages/LandingPage/Content';
+import Content from 'pages/landing/Content';
 
 const Landing = () => {
   const [position, setPostion] = useState(0);
