+++ src/pages/register/GameInput.tsx
