+++ src/pages/register/Games.tsx
