+++ src/pages/register/SetFavGame.tsx
