+++ src/pages/register/Terms.data.tsx
