+++ src/pages/register/Terms.tsx
